@@ -1,73 +1,58 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using System.Transactions;
 using Shuttle.Core.Contract;
 
-namespace Shuttle.Recall.Tests.Memory.Fakes
+namespace Shuttle.Recall.Tests.Memory.Fakes;
+
+public class MemoryPrimitiveEventRepository : IPrimitiveEventRepository
 {
-    public class MemoryPrimitiveEventRepository : IPrimitiveEventRepository
+    private readonly IPrimitiveEventStore _primitiveEventStore;
+    
+
+    public MemoryPrimitiveEventRepository(IPrimitiveEventStore primitiveEventStore)
     {
-        private static long _sequenceNumber = 1;
-        private readonly Dictionary<Guid, List<PrimitiveEvent>> _store;
+        _primitiveEventStore = Guard.AgainstNull(primitiveEventStore);
+    }
 
-        public MemoryPrimitiveEventRepository(Dictionary<Guid, List<PrimitiveEvent>> store)
+    public async Task RemoveAsync(Guid id)
+    {
+        await _primitiveEventStore.RemoveAggregateAsync(id);
+    }
+
+    public async ValueTask<long> SaveAsync(IEnumerable<PrimitiveEvent> primitiveEvents)
+    {
+        long sequenceNumber = 0;
+
+        var primitiveEventJournals = primitiveEvents.Select(item=>new PrimitiveEventJournal(item)).ToList();
+
+        foreach (var primitiveEventJournal in primitiveEventJournals)
         {
-            Guard.AgainstNull(store, nameof(store));
-
-            _store = store;
+            sequenceNumber = await _primitiveEventStore.AddAsync(primitiveEventJournal);
         }
 
-        public void Remove(Guid id)
+        if (Transaction.Current != null)
         {
-            _store.Remove(id);
+            Transaction.Current.EnlistVolatile(new PrimitiveEventJournalResourceManager(_primitiveEventStore, primitiveEventJournals), EnlistmentOptions.None);
         }
 
-        public IEnumerable<PrimitiveEvent> Get(Guid id)
-        {
-            return _store.ContainsKey(id) ? _store[id] : new List<PrimitiveEvent>();
-        }
+        return sequenceNumber;
+    }
 
-        public long Save(PrimitiveEvent primitiveEvent)
-        {
-            Guard.AgainstNull(primitiveEvent, nameof(primitiveEvent));
+    public async ValueTask<long> GetMaxSequenceNumberAsync()
+    {
+        return await _primitiveEventStore.GetMaxSequenceNumberAsync();
+    }
 
-            if (!_store.ContainsKey(primitiveEvent.Id))
-            {
-                _store.Add(primitiveEvent.Id, new List<PrimitiveEvent>());
-            }
+    public async Task<IEnumerable<PrimitiveEvent>> GetAsync(Guid id)
+    {
+        return await _primitiveEventStore.GetAsync(id);
+    }
 
-            primitiveEvent.SequenceNumber = _sequenceNumber++;
-
-            _store[primitiveEvent.Id].Add(primitiveEvent);
-
-            return _sequenceNumber;
-        }
-
-        public long GetSequenceNumber(Guid id)
-        {
-            return _sequenceNumber;
-        }
-
-        public async Task RemoveAsync(Guid id)
-        {
-            Remove(id);
-
-            await Task.CompletedTask;
-        }
-
-        public async Task<IEnumerable<PrimitiveEvent>> GetAsync(Guid id)
-        {
-            return await Task.FromResult(Get(id));
-        }
-
-        public async ValueTask<long> SaveAsync(PrimitiveEvent primitiveEvent)
-        {
-            return await new ValueTask<long>(Save(primitiveEvent));
-        }
-
-        public async ValueTask<long> GetSequenceNumberAsync(Guid id)
-        {
-            return await new ValueTask<long>(GetSequenceNumber(id));
-        }
+    public async ValueTask<long> GetSequenceNumberAsync(Guid id)
+    {
+        return await _primitiveEventStore.GetSequenceNumberAsync(id);
     }
 }
