@@ -1,24 +1,23 @@
 ﻿using Microsoft.Extensions.Logging;
 
-namespace Shuttle.Recall.Tests
+namespace Shuttle.Recall.Tests;
+
+public class FixtureFileLoggerProvider : ILoggerProvider
 {
-    public class FixtureFileLoggerProvider : ILoggerProvider
+    private readonly FixtureFileLogger _logger;
+
+    public FixtureFileLoggerProvider(string name)
     {
-        private readonly FixtureFileLogger _logger;
+        _logger = new(name);
+    }
 
-        public FixtureFileLoggerProvider(string name)
-        {
-            _logger = new FixtureFileLogger(name);
-        }
+    public void Dispose()
+    {
+        _logger.Dispose();
+    }
 
-        public void Dispose()
-        {
-            _logger.Dispose();
-        }
-
-        public ILogger CreateLogger(string categoryName)
-        {
-            return _logger;
-        }
+    public ILogger CreateLogger(string categoryName)
+    {
+        return _logger;
     }
 }
