@@ -1,14 +1,13 @@
-﻿namespace Shuttle.Recall.Tests
-{
-	public class OrderItem
-	{
-		public string Product { get; set; }
-		public double Quantity { get; set; }
-		public double Cost { get; set; }
+﻿namespace Shuttle.Recall.Tests;
 
-		public double Total()
-		{
-			return Quantity*Cost;
-		}
-	}
+public class OrderItem
+{
+    public double Cost { get; set; }
+    public string Product { get; set; } = string.Empty;
+    public double Quantity { get; set; }
+
+    public double Total()
+    {
+        return Quantity * Cost;
+    }
 }
