@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace Shuttle.Recall.Tests
-{
-	public class InvalidStatusChangeException : Exception
-	{
-		public InvalidStatusChangeException(string message) : base(message)
-		{
-		}
+namespace Shuttle.Recall.Tests;
 
-		public InvalidStatusChangeException()
-		{
-		}
-	}
+public class InvalidStatusChangeException : Exception
+{
+    public InvalidStatusChangeException(string message) : base(message)
+    {
+    }
+
+    public InvalidStatusChangeException()
+    {
+    }
 }
