@@ -1,107 +1,104 @@
 ﻿using System;
 using Shuttle.Core.Contract;
 
-namespace Shuttle.Recall.Tests
+namespace Shuttle.Recall.Tests;
+
+public enum OrderProcessStatus
 {
-    public enum OrderProcessStatus
+    Open = 0,
+    Picking = 1,
+    Fulfilled = 2,
+    Cancelled = 3
+}
+
+public class OrderProcess
+{
+    public OrderProcess(Guid id)
     {
-        Open = 0,
-        Picking = 1,
-        Fulfilled = 2,
-        Cancelled = 3
+        Id = id;
     }
 
-    public class OrderProcess
+    public Guid Id { get; }
+
+    public OrderProcessStatus Status { get; private set; }
+
+    public StatusChanged Cancel()
     {
-        public OrderProcess(Guid id)
+        var result = new StatusChanged
         {
-            Id = id;
-        }
+            Status = OrderProcessStatus.Cancelled
+        };
 
-        public Guid Id { get; }
+        On(result);
 
-        public OrderProcessStatus Status { get; private set; }
+        return result;
+    }
 
-        public StatusChanged Cancel()
+    public bool CanChangeStatusTo(OrderProcessStatus status)
+    {
+        switch (status)
         {
-            var result = new StatusChanged
+            case OrderProcessStatus.Open:
             {
-                Status = OrderProcessStatus.Cancelled
-            };
-
-            On(result);
-
-            return result;
-        }
-
-        public StatusChanged StartPicking()
-        {
-            var result = new StatusChanged
-            {
-                Status = OrderProcessStatus.Picking
-            };
-
-            On(result);
-
-            return result;
-        }
-
-        public StatusChanged Fulfill()
-        {
-            var result = new StatusChanged
-            {
-                Status = OrderProcessStatus.Fulfilled
-            };
-
-            On(result);
-
-            return result;
-        }
-
-        private void On(StatusChanged statusChanged)
-        {
-            Guard.AgainstNull(statusChanged, nameof(statusChanged));
-
-            InvariantStatus(statusChanged.Status);
-
-            Status = statusChanged.Status;
-        }
-
-        private void InvariantStatus(OrderProcessStatus status)
-        {
-            if (CanChangeStatusTo(status))
-            {
-                return;
+                return false;
             }
-
-            throw new InvalidStatusChangeException(status == OrderProcessStatus.Open
-                ? "An order can never be placed into the 'Open' status from any other status."
-                : $"The status cannot be changed to '{status}' from '{Status}'");
-        }
-
-        public bool CanChangeStatusTo(OrderProcessStatus status)
-        {
-            switch (status)
+            case OrderProcessStatus.Picking:
             {
-                case OrderProcessStatus.Open:
-                {
-                    return false;
-                }
-                case OrderProcessStatus.Picking:
-                {
-                    return Status == OrderProcessStatus.Open;
-                }
-                case OrderProcessStatus.Fulfilled:
-                {
-                    return Status == OrderProcessStatus.Picking;
-                }
-                case OrderProcessStatus.Cancelled:
-                {
-                    return Status == OrderProcessStatus.Open;
-                }
+                return Status == OrderProcessStatus.Open;
             }
-
-            return true;
+            case OrderProcessStatus.Fulfilled:
+            {
+                return Status == OrderProcessStatus.Picking;
+            }
+            case OrderProcessStatus.Cancelled:
+            {
+                return Status == OrderProcessStatus.Open;
+            }
         }
+
+        return true;
+    }
+
+    public StatusChanged Fulfill()
+    {
+        var result = new StatusChanged
+        {
+            Status = OrderProcessStatus.Fulfilled
+        };
+
+        On(result);
+
+        return result;
+    }
+
+    private void InvariantStatus(OrderProcessStatus status)
+    {
+        if (CanChangeStatusTo(status))
+        {
+            return;
+        }
+
+        throw new InvalidStatusChangeException(status == OrderProcessStatus.Open
+            ? "An order can never be placed into the 'Open' status from any other status."
+            : $"The status cannot be changed to '{status}' from '{Status}'");
+    }
+
+    private void On(StatusChanged statusChanged)
+    {
+        InvariantStatus(Guard.AgainstNull(statusChanged).Status);
+
+        Status = statusChanged.Status;
+    }
+
+    public StatusChanged StartPicking()
+    {
+        var result = new StatusChanged
+        {
+            Status = OrderProcessStatus.Picking
+        };
+
+        On(result);
+
+        return result;
     }
 }
