@@ -1,7 +1,6 @@
-﻿namespace Shuttle.Recall.Tests
+﻿namespace Shuttle.Recall.Tests;
+
+public class StatusChanged
 {
-	public class StatusChanged
-	{
-		public OrderProcessStatus Status { get; set; }
-	}
+    public OrderProcessStatus Status { get; set; }
 }
