@@ -6,92 +6,53 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Shuttle.Core.Contract;
 
-namespace Shuttle.Recall.Tests
+namespace Shuttle.Recall.Tests;
+
+public static class ServiceProviderExtensions
 {
-    public static class ServiceProviderExtensions
+    public static ILogger<T> GetLogger<T>(this IServiceProvider serviceProvider)
     {
-        public static IServiceProvider StartHostedServices(this IServiceProvider serviceProvider)
+        return Guard.AgainstNull(serviceProvider).GetRequiredService<ILoggerFactory>().CreateLogger<T>();
+    }
+
+    public static ILogger GetLogger(this IServiceProvider serviceProvider)
+    {
+        return Guard.AgainstNull(serviceProvider).GetRequiredService<ILoggerFactory>().CreateLogger("Fixture");
+    }
+
+    public static async Task<IServiceProvider> StartHostedServicesAsync(this IServiceProvider serviceProvider)
+    {
+        var logger = Guard.AgainstNull(serviceProvider).GetLogger();
+
+        logger.LogInformation("[StartHostedServices]");
+
+        foreach (var hostedService in serviceProvider.GetServices<IHostedService>())
         {
-            return StartHostedServicesAsync(serviceProvider, true).GetAwaiter().GetResult();
+            logger.LogInformation($"[HostedService-starting] : {hostedService.GetType().Name}");
+
+            await hostedService.StartAsync(CancellationToken.None).ConfigureAwait(false);
+
+            logger.LogInformation($"[HostedService-started] : {hostedService.GetType().Name}");
         }
 
-        public static IServiceProvider StopHostedServices(this IServiceProvider serviceProvider)
+        return serviceProvider;
+    }
+
+    public static async Task<IServiceProvider> StopHostedServicesAsync(this IServiceProvider serviceProvider)
+    {
+        var logger = Guard.AgainstNull(serviceProvider).GetLogger();
+
+        logger.LogInformation("[StopHostedServices]");
+
+        foreach (var hostedService in serviceProvider.GetServices<IHostedService>())
         {
-            return StopHostedServicesAsync(serviceProvider, true).GetAwaiter().GetResult();
+            logger.LogInformation($"[HostedService-stopping] : {hostedService.GetType().Name}");
+
+            await hostedService.StopAsync(CancellationToken.None).ConfigureAwait(false);
+
+            logger.LogInformation($"[HostedService-stopped] : {hostedService.GetType().Name}");
         }
 
-        public static async Task<IServiceProvider> StartHostedServicesAsync(this IServiceProvider serviceProvider)
-        {
-            return await StartHostedServicesAsync(serviceProvider, false).ConfigureAwait(false);
-        }
-
-        public static async Task<IServiceProvider> StopHostedServicesAsync(this IServiceProvider serviceProvider)
-        {
-            return await StopHostedServicesAsync(serviceProvider, false).ConfigureAwait(false);
-        }
-
-        private static async Task<IServiceProvider> StartHostedServicesAsync(IServiceProvider serviceProvider, bool sync)
-        {
-            Guard.AgainstNull(serviceProvider, nameof(serviceProvider));
-
-            var logger = serviceProvider.GetLogger();
-
-            logger.LogInformation($"[StartHostedServices]");
-
-            foreach (var hostedService in serviceProvider.GetServices<IHostedService>())
-            {
-                logger.LogInformation($"[HostedService-starting] : {hostedService.GetType().Name}");
-
-                if (sync)
-                {
-                    hostedService.StartAsync(CancellationToken.None).GetAwaiter().GetResult();
-                }
-                else
-                {
-                    await hostedService.StartAsync(CancellationToken.None).ConfigureAwait(false);
-                }
-
-                logger.LogInformation($"[HostedService-started] : {hostedService.GetType().Name}");
-            }
-
-            return serviceProvider;
-        }
-
-        private static async Task<IServiceProvider> StopHostedServicesAsync(IServiceProvider serviceProvider, bool sync)
-        {
-            Guard.AgainstNull(serviceProvider, nameof(serviceProvider));
-
-            var logger = serviceProvider.GetLogger();
-
-            logger.LogInformation($"[StopHostedServices]");
-            
-            foreach (var hostedService in serviceProvider.GetServices<IHostedService>())
-            {
-                logger.LogInformation($"[HostedService-stopping] : {hostedService.GetType().Name}");
-
-                if (sync)
-                {
-                    hostedService.StopAsync(CancellationToken.None).GetAwaiter().GetResult();
-                }
-                else
-                {
-                    await hostedService.StopAsync(CancellationToken.None).ConfigureAwait(false);
-                }
-
-                logger.LogInformation($"[HostedService-stopped] : {hostedService.GetType().Name}");
-            }
-
-            return serviceProvider;
-        }
-
-        public static ILogger<T> GetLogger<T>(this IServiceProvider serviceProvider)
-        {
-            return Guard.AgainstNull(serviceProvider, nameof(serviceProvider)).GetRequiredService<ILoggerFactory>().CreateLogger<T>();
-        }
-
-        public static ILogger GetLogger(this IServiceProvider serviceProvider)
-        {
-            return Guard.AgainstNull(serviceProvider, nameof(serviceProvider)).GetRequiredService<ILoggerFactory>().CreateLogger("Fixture");
-        }
+        return serviceProvider;
     }
 }
