@@ -1,22 +1,11 @@
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-#if NETFRAMEWORK
-[assembly: AssemblyTitle(".NET Framework")]
-#endif
-
-#if NETCOREAPP
-[assembly: AssemblyTitle(".NET Core")]
-#endif
-
-#if NETSTANDARD
-[assembly: AssemblyTitle(".NET Standard")]
-#endif
-
-[assembly: AssemblyVersion("18.0.0.0")]
-[assembly: AssemblyCopyright("Copyright (c) 2024, Eben Roux")]
+[assembly: AssemblyTitle(".NET Unified Platform")]
+[assembly: AssemblyVersion("20.0.0.0")]
+[assembly: AssemblyCopyright("Copyright (c) 2025, Eben Roux")]
 [assembly: AssemblyProduct("Shuttle.Recall.Tests")]
 [assembly: AssemblyCompany("Eben Roux")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyInformationalVersion("18.0.0")]
+[assembly: AssemblyInformationalVersion("20.0.0")]
 [assembly: ComVisible(false)]
